requirements.txt
djangorestframework==3.12.4
django==3.2.6
djangorestframework
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
django-elasticsearch-dsl==7.2.0